# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aegis API",
    "description": "Orquestador del agente financiero: propuestas, políticas, Guardian y auditoría.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/proposals": {
      "get": {
        "summary": "Lista las propuestas del usuario, de la más reciente a la más antigua",
        "tags": [
          "proposals"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "proposals": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "userId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DRAFT",
                              "POLICY_CHECK",
                              "GUARDIAN_REVIEW",
                              "PENDING_USER",
                              "AUTO_APPROVED",
                              "SIGNED",
                              "SUBMITTED",
                              "CONFIRMED",
                              "REJECTED",
                              "DENIED",
                              "EXPIRED",
                              "FAILED"
                            ]
                          },
                          "summary": {
                            "type": "string",
                            "maxLength": 280
                          },
                          "actions": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "enum": [
                                    "PAYMENT"
                                  ]
                                },
                                "destinationId": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "asset": {
                                  "type": "string",
                                  "enum": [
                                    "XLM",
                                    "USDC_TEST"
                                  ]
                                },
                                "amount": {
                                  "type": "string",
                                  "pattern": "^\\d+(\\.\\d{1,7})?$"
                                },
                                "memo": {
                                  "type": "string",
                                  "maxLength": 28,
                                  "nullable": true
                                },
                                "label": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 64
                                }
                              },
                              "required": [
                                "type",
                                "destinationId",
                                "asset",
                                "amount",
                                "label"
                              ],
                              "additionalProperties": false
                            },
                            "minItems": 1,
                            "maxItems": 10
                          },
                          "policy": {
                            "type": "object",
                            "properties": {
                              "decision": {
                                "type": "string",
                                "enum": [
                                  "AUTO_APPROVE",
                                  "REQUIRE_USER",
                                  "DENY"
                                ]
                              },
                              "reasons": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "ruleId": {
                                      "type": "string",
                                      "enum": [
                                        "P-01",
                                        "P-02",
                                        "P-03",
                                        "P-04",
                                        "P-05",
                                        "P-06",
                                        "P-07",
                                        "P-08",
                                        "P-09"
                                      ]
                                    },
                                    "effect": {
                                      "type": "string",
                                      "enum": [
                                        "AUTO_APPROVE",
                                        "REQUIRE_USER",
                                        "DENY"
                                      ]
                                    },
                                    "message": {
                                      "type": "string"
                                    },
                                    "actionIndex": {
                                      "type": "integer",
                                      "minimum": 0
                                    }
                                  },
                                  "required": [
                                    "ruleId",
                                    "effect",
                                    "message"
                                  ],
                                  "additionalProperties": false
                                }
                              },
                              "evaluatedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "decision",
                              "reasons",
                              "evaluatedAt"
                            ],
                            "additionalProperties": false,
                            "nullable": true
                          },
                          "risk": {
                            "type": "object",
                            "properties": {
                              "score": {
                                "type": "number",
                                "minimum": 0,
                                "maximum": 100
                              },
                              "level": {
                                "type": "string",
                                "enum": [
                                  "LOW",
                                  "MEDIUM",
                                  "HIGH",
                                  "CRITICAL"
                                ]
                              },
                              "signals": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "type": "string",
                                      "enum": [
                                        "G-01",
                                        "G-02",
                                        "G-03",
                                        "G-04",
                                        "G-05",
                                        "G-06",
                                        "G-07",
                                        "G-08",
                                        "G-09"
                                      ]
                                    },
                                    "severity": {
                                      "type": "string",
                                      "enum": [
                                        "INFO",
                                        "WARN",
                                        "HIGH"
                                      ]
                                    },
                                    "actionIndex": {
                                      "type": "integer",
                                      "minimum": 0
                                    },
                                    "data": {
                                      "type": "object",
                                      "additionalProperties": {}
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "severity",
                                    "data"
                                  ],
                                  "additionalProperties": false
                                }
                              },
                              "balanceAfter": {
                                "type": "string",
                                "pattern": "^\\d+(\\.\\d{1,7})?$"
                              },
                              "evaluatedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "score",
                              "level",
                              "signals",
                              "balanceAfter",
                              "evaluatedAt"
                            ],
                            "additionalProperties": false,
                            "nullable": true
                          },
                          "explanation": {
                            "type": "object",
                            "properties": {
                              "summary": {
                                "type": "string"
                              },
                              "warnings": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "signalId": {
                                      "type": "string",
                                      "enum": [
                                        "G-01",
                                        "G-02",
                                        "G-03",
                                        "G-04",
                                        "G-05",
                                        "G-06",
                                        "G-07",
                                        "G-08",
                                        "G-09"
                                      ]
                                    },
                                    "text": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "signalId",
                                    "text"
                                  ],
                                  "additionalProperties": false
                                }
                              },
                              "generatedBy": {
                                "type": "string",
                                "enum": [
                                  "llm",
                                  "template"
                                ]
                              }
                            },
                            "required": [
                              "summary",
                              "warnings",
                              "generatedBy"
                            ],
                            "additionalProperties": false,
                            "nullable": true
                          },
                          "unsignedXdr": {
                            "type": "string",
                            "nullable": true
                          },
                          "txHash": {
                            "type": "string",
                            "nullable": true
                          },
                          "failureReason": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "status",
                          "summary",
                          "actions",
                          "createdAt",
                          "expiresAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "proposals"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ]
}